Move bbox_iou to models and remove unused codes
